+++ iNavTelemetryOSX/iNavTelemetryOSXApp.swift
@@ -1,0 +1,17 @@
+//
+//  iNavTelemetryOSXApp.swift
+//  iNavTelemetryOSX
+//
+//  Created by Bosko Petreski on 9/26/21.
+//
+
+import SwiftUI
+
+@main
+struct iNavTelemetryOSXApp: App {
+    var body: some Scene {
+        WindowGroup {
+            MainView()
+        }
+    }
+}
